pyproject.toml
[project]
name = "13-llm-ans-compression"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.35.3",
    "llama-cpp-python>=0.3.16",
    "numpy>=2.3.3",
]

[[tool.uv.index]]
url = "https://abetlen.github.io/llama-cpp-python/whl/metal"
